Adding students to a course
The add student button now adds all selected students to the current
course thats being edited and all of the courses part evaluations. I've
also added a search function that allows the user to search for students
to add to a course. I will add another one like this for course parts.